Fix disabled tests
